LRCI-3385 Change Testray Project IDs to longs instead of ints

+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/BaseTestrayServer.java
@@ -338,7 +338,7 @@ public abstract class BaseTestrayServer implements TestrayServer {
 
 	private JenkinsResultsParserUtil.HTTPAuthorization _httpAuthorization;
 	private Map<String, TestrayCaseType> _testrayCaseTypes;
-	private Map<Integer, TestrayProject> _testrayProjectsByID;
+	private Map<Long, TestrayProject> _testrayProjectsByID;
 	private Map<String, TestrayProject> _testrayProjectsByName;
 	private final URL _url;
