+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/TestrayProject.java
@@ -146,7 +146,7 @@ public class TestrayProject {
 		return _jsonObject.getString("description");
 	}
 
-	public int getID() {
+	public long getID() {
 		return _jsonObject.getInt("testrayProjectId");
 	}
